implement hw 1
